Added authorization to the AnnotationController

+++ src/main/java/com/certh/spring/data/mongodb/controller/AnnotationController.java
@@ -1,13 +1,16 @@
 package com.certh.spring.data.mongodb.controller;
 
-import com.certh.spring.data.mongodb.model.Annotation;
+import com.certh.spring.data.mongodb.exception.ResourceNotFoundException;
+import com.certh.spring.data.mongodb.exception.ResourceOwnershipException;
+import com.certh.spring.data.mongodb.models.Annotation;
 import com.certh.spring.data.mongodb.repository.AnnotationRepository;
+import com.certh.spring.data.mongodb.security.services.AnnotationService;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,20 +27,31 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api")
 public class AnnotationController {
 
+  private final AnnotationRepository annotationRepository;
+  private final AnnotationService annotationService;
+
   @Autowired
-  AnnotationRepository annotationRepository;
+  public AnnotationController(AnnotationRepository annotationRepository,
+      AnnotationService annotationService) {
+    this.annotationRepository = annotationRepository;
+    this.annotationService = annotationService;
+  }
 
   @GetMapping("/annotations")
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
   public ResponseEntity<List<Annotation>> getAllAnnotations(
-      @RequestParam(required = false) String videoId,
-      @RequestParam(required = false) String userId) {
-    try {
-      List<Annotation> annotations = new ArrayList<Annotation>();
+      @RequestParam() String videoId) {
 
-      if (videoId == null && userId == null) {
-        annotationRepository.findAll().forEach(annotations::add);
-      } else {
-        annotationRepository.findByVideoIdAndUserId(videoId, userId).forEach(annotations::add);
+    String userId = annotationService.findLoggedInUser();
+
+    try {
+      List<Annotation> annotations = new ArrayList<>();
+
+      Iterable<Annotation> iterable = annotationRepository.findByVideoIdAndUserId(videoId,
+          userId);
+
+      for (Annotation annotation : iterable) {
+        annotations.add(annotation);
       }
 
       if (annotations.isEmpty()) {
@@ -51,52 +65,71 @@ public class AnnotationController {
   }
 
   @GetMapping("/annotations/{id}")
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
   public ResponseEntity<Annotation> getAnnotationById(@PathVariable("id") String id) {
-    Optional<Annotation> annotationData = annotationRepository.findById(id);
+    String userId = annotationService.findLoggedInUser();
 
-    if (annotationData.isPresent()) {
-      return new ResponseEntity<>(annotationData.get(), HttpStatus.OK);
-    } else {
-      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    Annotation _annotation = annotationRepository.findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("Not found Annotation with id = " + id));
+
+    if (!_annotation.getUserId().contentEquals(userId)) {
+      throw new ResourceOwnershipException(
+          "You do not have permission to access Annotation with id = " + id);
     }
+
+    return new ResponseEntity<>(_annotation, HttpStatus.OK);
   }
 
   @PostMapping("/annotations")
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
   public ResponseEntity<Annotation> createAnnotation(@RequestBody Annotation annotation) {
     try {
-      Annotation _annotations = annotationRepository.save(
-          new Annotation(annotation.getRectangle(), annotation.getFrameNumber(),
-              annotation.getSecond(), annotation.getDescription(), annotation.getDropdownValue(),
-              annotation.getVideoId(), annotation.getUserId()));
-      return new ResponseEntity<>(_annotations, HttpStatus.CREATED);
+      Annotation _annotation = annotationService.createAnnotation(annotation);
+      _annotation = annotationRepository.save(_annotation);
+      return new ResponseEntity<>(_annotation, HttpStatus.CREATED);
     } catch (Exception e) {
       return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
   @PutMapping("/annotations/{id}")
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
   public ResponseEntity<Annotation> updateAnnotation(@PathVariable("id") String id,
       @RequestBody Annotation annotation) {
-    Optional<Annotation> annotationData = annotationRepository.findById(id);
+    String userId = annotationService.findLoggedInUser();
 
-    if (annotationData.isPresent()) {
-      Annotation _annotation = annotationData.get();
-      _annotation.setRectangle(annotation.getRectangle());
-      _annotation.setFrameNumber(annotation.getFrameNumber());
-      _annotation.setSecond(annotation.getSecond());
-      _annotation.setDescription(annotation.getDescription());
-      _annotation.setDropdownValue(annotation.getDropdownValue());
-      _annotation.setVideoId(annotation.getVideoId());
-      _annotation.setUserId(annotation.getUserId());
+    Annotation _annotation = annotationRepository.findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("Not found Annotation with id = " + id));
 
-      return new ResponseEntity<>(annotationRepository.save(_annotation), HttpStatus.OK);
-    } else {
-      return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    if (!_annotation.getUserId().contentEquals(userId)) {
+      throw new ResourceOwnershipException(
+          "You do not have permission to modify Annotation with id = " + id);
     }
+
+    _annotation.setRectangle(annotation.getRectangle());
+    _annotation.setFrameNumber(annotation.getFrameNumber());
+    _annotation.setSecond(annotation.getSecond());
+    _annotation.setDescription(annotation.getDescription());
+    _annotation.setDropdownValue(annotation.getDropdownValue());
+    _annotation.setVideoId(annotation.getVideoId());
+    _annotation.setUserId(userId);
+
+    return new ResponseEntity<>(annotationRepository.save(_annotation), HttpStatus.OK);
   }
 
   @DeleteMapping("/annotations/{id}")
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
   public ResponseEntity<HttpStatus> deleteAnnotation(@PathVariable("id") String id) {
+    String userId = annotationService.findLoggedInUser();
+
+    Annotation _annotation = annotationRepository.findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("Not found Annotation with id = " + id));
+
+    if (!_annotation.getUserId().contentEquals(userId)) {
+      throw new ResourceOwnershipException(
+          "You do not have permission to modify Annotation with id = " + id);
+    }
+
     try {
       annotationRepository.deleteById(id);
       return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -106,9 +139,16 @@ public class AnnotationController {
   }
 
   @DeleteMapping("/annotations")
-  public ResponseEntity<HttpStatus> deleteAllAnnotations() {
+  @PreAuthorize("hasRole('TRAINER') or hasRole('TRAINEE') or hasRole('ADMIN')")
+  public ResponseEntity<HttpStatus> deleteAllAnnotations(@RequestParam(required = false) String videoId) {
+    String userId = annotationService.findLoggedInUser();
+
     try {
-      annotationRepository.deleteAll();
+      if (videoId == null) {
+        annotationRepository.deleteByUserId(userId);
+      } else {
+        annotationRepository.deleteByUserIdAndVideoId(userId, videoId);
+      }
       return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     } catch (Exception e) {
       return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
